PC_CONTROL_CODE/requirements.txt
pySerial == 3.5